refactor: `KeywordCondition`, `KeywordFilter`의 패키지를 이동한다.
