+++ src/main/java/com/clova/anifriends/domain/recruitment/dto/request/KeywordFilter.java
@@ -1,6 +1,8 @@
-package com.clova.anifriends.domain.recruitment.controller;
+package com.clova.anifriends.domain.recruitment.dto.request;
 
 import com.clova.anifriends.domain.common.EnumType;
+import com.clova.anifriends.domain.recruitment.service.KeywordCondition;
+import com.clova.anifriends.domain.recruitment.service.KeywordConditionByShelter;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
