+++ src/main/java/com/clova/anifriends/domain/recruitment/service/KeywordCondition.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.recruitment.controller;
+package com.clova.anifriends.domain.recruitment.service;
 
 public record KeywordCondition(boolean titleFilter, boolean contentFilter, boolean shelterNameFilter) {
