+++ src/main/java/com/clova/anifriends/domain/recruitment/service/KeywordConditionByShelter.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.recruitment.controller;
+package com.clova.anifriends.domain.recruitment.service;
 
 public record KeywordConditionByShelter(boolean titleFilter, boolean contentFilter) {
