+++ src/main/java/com/clova/anifriends/domain/recruitment/service/RecruitmentCacheService.java
@@ -1,7 +1,6 @@
 package com.clova.anifriends.domain.recruitment.service;
 
 import com.clova.anifriends.domain.recruitment.Recruitment;
-import com.clova.anifriends.domain.recruitment.controller.KeywordCondition;
 import com.clova.anifriends.domain.recruitment.repository.RecruitmentCacheRepository;
 import com.clova.anifriends.domain.recruitment.repository.RecruitmentRepository;
 import java.util.List;
