+++ src/test/java/com/clova/anifriends/domain/recruitment/service/RecruitmentCacheServiceTest.java
@@ -6,7 +6,6 @@ import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.times;
 
 import com.clova.anifriends.domain.recruitment.Recruitment;
-import com.clova.anifriends.domain.recruitment.controller.KeywordCondition;
 import com.clova.anifriends.domain.recruitment.repository.RecruitmentRedisRepository;
 import com.clova.anifriends.domain.recruitment.repository.RecruitmentRepository;
 import com.clova.anifriends.domain.recruitment.support.fixture.RecruitmentFixture;
